Refactor project
Note: Skimping on proper commit frequency due to being in the project
initiation phase.

This commit changes the file structure to include a timestamp of the
files' creation. A symbolic link 'latest' is also added which links to
the latest indexed version. If possible, the memento API will be added
on top of this structure in nginx (extensions) to make standardized time
travel possible.